http://trac.atomiclabs.com/ticket/285
Fixed bug with HTTPProtocol not resetting the parsers which was causing 
it to copy the first event in a connection repeatedly for every request
rather than generating unique events for each requests within a 
pipelined connection.

http://trac.atomiclabs.com/ticket/286

Fixed bug in LogOutputReactor that was causing it to delete files when 
it shouldn't

+++ platform/reactors/LogOutputReactor.cpp
@@ -122,7 +122,7 @@ void LogOutputReactor::stop(void)
 			m_codec_ptr->finish(m_log_stream);
 			m_log_stream.close();
 			// remove the log file if no events were written to it
-			if (getEventsOut() == 0) {
+			if (getEventsOut() == 0 && boost::filesystem::file_size(m_log_filename) == 0) {
 				boost::filesystem::remove(m_log_filename);
 				PION_LOG_DEBUG(m_logger, "Closing empty output log (removing file): " << m_log_filename);
 			} else {
